Renamed model to protocol to match the original endpoint
