+++ vllm/entrypoints/sync_openai/protocol.py
